screw user friendly
i've given up on making it user friendly.

if you want what i'll be using. it's live_gigaspeech_dictation_v20.py it's the most stable i've ever gotten it. if it doesn't work and you have all the files. it's probably your device_index. that's your mic. set it to the right one.

for future reference. 31.3 is how i want the CLI to look. and pygui_shared_v10 is how i want the main GUI to look. and waveform_test_v2 is how i want the waveform to look like. the rest are just features that i doubt will ever be implemented. this is just the bare bones of what i wanted.

I WILL GIVE A SHOUT OUT/KUDOS TO ANYONE WHO FIXES IT. your name will be in the readme file as someone who merged or integrated the files together.