Update live_gigaspeech_dictation_v20.py
the other version i messed with words to number and it broke it. this one works.

you can use words to number 7 or 8. 7 adds your numbers if said in digits. 8 doesn't add. but has it's own limits. 8 is better unless you want your thing to add for you. 123 it will output 6. but 8 will output 123.

TL;DR version 7 adds version 8 gives it as is.